backend/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
sqlalchemy==2.0.48
pymysql==1.1.2
pydantic==2.12.5
pydantic-settings==2.13.1
passlib==1.7.4
bcrypt==3.2.0
python-multipart==0.0.22
openai==2.29.0
python-dotenv==1.2.2
numpy==2.4.3
cryptography==42.0.5
openai-whisper==20250625
torch==2.11.0
librosa==0.10.2
soundfile==0.12.1

